Created settings page with add/delete user.
When logged in, the Settings link under the account takes the user into the settings page. From here, they can add a user to the database or delete one from a dropdown list that is automatically populated from the database. The footer has also been modularized so that it can be included in every page. It must be required and then can be displayed by calling displayFooter(). 
I have also commented out a test in Salt_n_Pepper.php so that it doesn't display on every page.